+++ src/MarketCurves/MarketCurves.Query.Service/Features/GetMarketCurveDetail/PointDto.cs
@@ -8,7 +8,7 @@ namespace MarketCurves.Query.Service.Features.GetMarketCurveDetail
         public string Tenor { get; set; }
         public string Vendor { get; set; }
         public string Name { get; set; }
-        public short DateLag { get; set; }
+        public int DateLag { get; set; }
         public bool IsMandatory { get; set; }
         public string PriceType { get; set; }
     }
